backend/requirements.txt
fastapi
uvicorn
PyPDF2
anthropic
python-dotenv
